# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Meta API",
    "description": "Meta swagger api",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apikeyAuth": {
      "type": "apiKey",
      "in": "headers",
      "name": "authorization",
      "description": "請加上api key"
    }
  },
  "paths": {
    "/posts/": {
      "get": {
        "tags": [
          "Posts-貼文"
        ],
        "summary": "取得所有貼文",
        "description": "取得所有貼文",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "description": "查詢貼文官內容關鍵字",
            "required": false,
            "type": "string"
          },
          {
            "name": "size",
            "in": "query",
            "description": "查詢貼文數量，上限 50 筆，預設 50 筆",
            "required": false,
            "type": "integer",
            "maximum": 50
          },
          {
            "name": "timeSort",
            "in": "query",
            "description": "發文時間排序方式，預設 desc 降冪排序，asc=遠到近，desc=近到遠",
            "required": false,
            "type": "string",
            "enum": [
              "asc",
              "desc"
            ]
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Posts-貼文"
        ],
        "summary": "新增貼文",
        "description": "新增貼文",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  }
}